requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow>=10.0.0
pillow-heif>=0.13.0
gunicorn==21.2.0
cairosvg>=2.7.0
pypdf>=4.0.0
PyMuPDF>=1.23.0
reportlab>=4.0.0
pikepdf>=8.0.0
mammoth>=1.6.0
weasyprint>=60.0
rembg>=2.0.0
pytesseract>=0.3.10
opencv-python-headless>=4.8.0
gradio-client>=0.16.0
mediapipe>=0.10.0
scipy>=1.11.0

